requirements.txt
flask==3.0.0
requests==2.31.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pyinstaller==6.3.0


